rename the modal task component to addTask
