+++ src/assets/styles/TaskBar.scss
@@ -189,7 +189,7 @@
     }
   }
 
-  .ModalTask {
+  .AddTask {
     width: 25vw;
     height: 40vh;
     display: flex;
